Fix appscript column value mapping
Boolean values weren't working in the graphs, so now they get casted to numbers. Also, the mappings were still set to the test mappings. The ones for the competition are now set (DATE, HEIGHT, COMMENTS). I should make a way for these mappings to be defined per-competition instead of being defined directly in the UI Sheet, but this will do for now.