Update appscripts and Scouting UI column mapping
Column mapping is now read from the specified data spreadsheet's constants sheet. Previously, when changing out which data sheet was being read, you would have to go into the appscripts code to manually change the column mapping functions. Now, the code for the function is placed in the constants sheet under the column which the function will apply to. Switching data speadsheets now is fully automatic.